refactor(packaging): update package structure for improved design clarity
